requirements.txt
requests
python-dotenv
selenium
webdriver-manager
fastmcp
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart
